Add function to get Player choice
Function getPlayerChoice makes prompt asking
for either Rock, Paper or Scissors.

After user type the choice, the function automatically
converts letters to lower case so it can match with
elements from global array.

There is a while loop so that if the user press cancel
on the prompt, it will just loop until something is
typed or the "ok" button is pressed.